+++ src/Key/Derivation/Scrypt.php
@@ -3,7 +3,7 @@
  * Zend Framework (http://framework.zend.com/)
  *
  * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2005-2016 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   http://framework.zend.com/license/new-bsd New BSD License
  */
 
@@ -51,7 +51,7 @@ abstract class Scrypt
 
         $s = '';
         for ($i = 0; $i < $p; $i++) {
-            $s .= self::scryptROMix(substr($b, $i * 128 * $r, 128 * $r), $n, $r);
+            $s .= self::scryptROMix(mb_substr($b, $i * 128 * $r, 128 * $r, '8bit'), $n, $r);
         }
 
         return Pbkdf2::calc('sha256', $password, $s, 1, $length);
@@ -92,16 +92,16 @@ abstract class Scrypt
      */
     protected static function scryptBlockMix($b, $r)
     {
-        $x    = substr($b, -64);
+        $x    = mb_substr($b, -64, null, '8bit');
         $even = '';
         $odd  = '';
         $len  = 2 * $r;
 
         for ($i = 0; $i < $len; $i++) {
             if (PHP_INT_SIZE === 4) {
-                $x = self::salsa208Core32($x ^ substr($b, 64 * $i, 64));
+                $x = self::salsa208Core32($x ^ mb_substr($b, 64 * $i, 64, '8bit'));
             } else {
-                $x = self::salsa208Core64($x ^ substr($b, 64 * $i, 64));
+                $x = self::salsa208Core64($x ^ mb_substr($b, 64 * $i, 64, '8bit'));
             }
             if ($i % 2 == 0) {
                 $even .= $x;
@@ -124,7 +124,7 @@ abstract class Scrypt
     {
         $b32 = [];
         for ($i = 0; $i < 16; $i++) {
-            list(, $b32[$i]) = unpack("V", substr($b, $i * 4, 4));
+            list(, $b32[$i]) = unpack("V", mb_substr($b, $i * 4, 4, '8bit'));
         }
 
         $x = $b32;
@@ -217,7 +217,7 @@ abstract class Scrypt
     {
         $b32 = [];
         for ($i = 0; $i < 16; $i++) {
-            list(, $b32[$i]) = unpack("V", substr($b, $i * 4, 4));
+            list(, $b32[$i]) = unpack("V", mb_substr($b, $i * 4, 4, '8bit'));
         }
 
         $x = $b32;
@@ -315,7 +315,7 @@ abstract class Scrypt
         if (PHP_INT_SIZE === 8) {
             $v = 'V';
         }
-        list(, $n) = unpack($v, substr($b, -64));
+        list(, $n) = unpack($v, mb_substr($b, -64, null, '8bit'));
         return $n;
     }
 }