+++ src/Password/Apache.php
@@ -3,7 +3,7 @@
  * Zend Framework (http://framework.zend.com/)
  *
  * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2005-2016 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   http://framework.zend.com/license/new-bsd New BSD License
  */
 
@@ -125,12 +125,12 @@ class Apache implements PasswordInterface
      */
     public function verify($password, $hash)
     {
-        if (substr($hash, 0, 5) === '{SHA}') {
+        if (mb_substr($hash, 0, 5, '8bit') === '{SHA}') {
             $hash2 = '{SHA}' . base64_encode(sha1($password, true));
             return Utils::compareStrings($hash, $hash2);
         }
 
-        if (substr($hash, 0, 6) === '$apr1$') {
+        if (mb_substr($hash, 0, 6, '8bit') === '$apr1$') {
             $token = explode('$', $hash);
             if (empty($token[2])) {
                 throw new Exception\InvalidArgumentException(
@@ -143,7 +143,7 @@ class Apache implements PasswordInterface
 
         $bcryptPattern = '/\$2[ay]?\$[0-9]{2}\$[' . addcslashes(static::BASE64, '+/') . '\.]{53}/';
 
-        if (strlen($hash) > 13 && ! preg_match($bcryptPattern, $hash)) { // digest
+        if (mb_strlen($hash, '8bit') > 13 && ! preg_match($bcryptPattern, $hash)) { // digest
             if (empty($this->userName) || empty($this->authName)) {
                 throw new Exception\RuntimeException(
                     'You must specify UserName and AuthName (realm) to verify the digest'
@@ -242,7 +242,7 @@ class Apache implements PasswordInterface
      */
     protected function toAlphabet64($value)
     {
-        return strtr(strrev(substr(base64_encode($value), 2)), self::BASE64, self::ALPHA64);
+        return strtr(strrev(mb_substr(base64_encode($value), 2, null, '8bit')), self::BASE64, self::ALPHA64);
     }
 
     /**
@@ -257,7 +257,7 @@ class Apache implements PasswordInterface
         if (null === $salt) {
             $salt = Rand::getString(8, self::ALPHA64);
         } else {
-            if (strlen($salt) !== 8) {
+            if (mb_strlen($salt, '8bit') !== 8) {
                 throw new Exception\InvalidArgumentException(
                     'The salt value for APR1 algorithm must be 8 characters long'
                 );
@@ -270,11 +270,11 @@ class Apache implements PasswordInterface
                 }
             }
         }
-        $len  = strlen($password);
+        $len  = mb_strlen($password, '8bit');
         $text = $password . '$apr1$' . $salt;
         $bin  = pack("H32", md5($password . $salt . $password));
         for ($i = $len; $i > 0; $i -= 16) {
-            $text .= substr($bin, 0, min(16, $i));
+            $text .= mb_substr($bin, 0, min(16, $i), '8bit');
         }
         for ($i = $len; $i > 0; $i >>= 1) {
             $text .= ($i & 1) ? chr(0) : $password[0];