+++ src/Symmetric/Padding/Pkcs7.php
@@ -3,7 +3,7 @@
  * Zend Framework (http://framework.zend.com/)
  *
  * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2005-2016 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   http://framework.zend.com/license/new-bsd New BSD License
  */
 
@@ -24,7 +24,7 @@ class Pkcs7 implements PaddingInterface
      */
     public function pad($string, $blockSize = 32)
     {
-        $pad = $blockSize - (strlen($string) % $blockSize);
+        $pad = $blockSize - (mb_strlen($string, '8bit') % $blockSize);
         return $string . str_repeat(chr($pad), $pad);
     }
 
@@ -37,11 +37,11 @@ class Pkcs7 implements PaddingInterface
      */
     public function strip($string)
     {
-        $end  = substr($string, -1);
+        $end  = mb_substr($string, -1, null, '8bit');
         $last = ord($end);
-        $len  = strlen($string) - $last;
-        if (substr($string, $len) == str_repeat($end, $last)) {
-            return substr($string, 0, $len);
+        $len  = mb_strlen($string, '8bit') - $last;
+        if (mb_substr($string, $len, null, '8bit') == str_repeat($end, $last)) {
+            return mb_substr($string, 0, $len, '8bit');
         }
         return false;
     }