+++ src/Utils.php
@@ -3,7 +3,7 @@
  * Zend Framework (http://framework.zend.com/)
  *
  * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2005-2016 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   http://framework.zend.com/license/new-bsd New BSD License
  */
 
@@ -36,8 +36,8 @@ class Utils
             return hash_equals($expected, $actual);
         }
 
-        $lenExpected  = strlen($expected);
-        $lenActual    = strlen($actual);
+        $lenExpected  = mb_strlen($expected, '8bit');
+        $lenActual    = mb_strlen($actual, '8bit');
         $len          = min($lenExpected, $lenActual);
 
         $result = 0;