+++ test/Key/Derivation/SaltedS2kTest.php
@@ -3,7 +3,7 @@
  * Zend Framework (http://framework.zend.com/)
  *
  * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2005-2016 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   http://framework.zend.com/license/new-bsd New BSD License
  */
 
@@ -41,8 +41,10 @@ class SaltedS2kTest extends \PHPUnit_Framework_TestCase
             $this->markTestSkipped('The mhash extension is not available');
             return;
         }
-        $this->setExpectedException('Zend\Crypt\Key\Derivation\Exception\InvalidArgumentException',
-                                    'The hash algorithm wrong is not supported by Zend\Crypt\Key\Derivation\SaltedS2k');
+        $this->setExpectedException(
+            'Zend\Crypt\Key\Derivation\Exception\InvalidArgumentException',
+            'The hash algorithm wrong is not supported by Zend\Crypt\Key\Derivation\SaltedS2k'
+        );
         $password = SaltedS2k::calc('wrong', 'test', $this->salt, 32);
     }
 
@@ -52,8 +54,10 @@ class SaltedS2kTest extends \PHPUnit_Framework_TestCase
             $this->markTestSkipped('The mhash extension is not available');
             return;
         }
-        $this->setExpectedException('Zend\Crypt\Key\Derivation\Exception\InvalidArgumentException',
-                                    'The salt size must be at least of 8 bytes');
+        $this->setExpectedException(
+            'Zend\Crypt\Key\Derivation\Exception\InvalidArgumentException',
+            'The salt size must be at least of 8 bytes'
+        );
         $password = SaltedS2k::calc('sha256', 'test', substr($this->salt, -1), 32);
     }
 }