+++ CHANGELOG.md
@@ -2,6 +2,59 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
+## 3.0.0 - TBD
+
+### Added
+
+- [#22](https://github.com/zendframework/zend-crypt/pull/22) adds a requirement
+  on `ext/mbstring` in order to install successfully.
+- [#25](https://github.com/zendframework/zend-crypt/pull/25) adds a new
+  symmetric encryption adapter for the OpenSSL extension; this is now the
+  default adapter used internally by the component when symmetric encryption is
+  required.
+- [#25](https://github.com/zendframework/zend-crypt/pull/25) adds support for
+  zend-math v3.
+- [#26](https://github.com/zendframework/zend-crypt/pull/26) adds
+  `Zend\Crypt\Password\Bcrypt::benchmarkCost()`, which allows you to find the
+  maximum cost value possible for your hardware within a 50ms timeframe.
+- [#11](https://github.com/zendframework/zend-crypt/pull/11) adds a new option
+  to the `Zend\Crypt\PublicKey\RsaOptions` class, `openssl_padding` (or
+  `setOpensslPadding()`; this is now consumed in
+  `Zend\Crypt\PublicKey\Rsa::encrypt()` and
+  `Zend\Crypt\PublicKey\Rsa::decrypt()`, instead of the optional `$padding`
+  argument.
+
+### Deprecated
+
+- [#25](https://github.com/zendframework/zend-crypt/pull/25) deprecates usage of the
+  mcrypt symmetric encryption adapter when used on PHP 7 versions, as PHP 7.1
+  will deprecate the mcrypt extension.
+
+### Removed
+
+- [#11](https://github.com/zendframework/zend-crypt/pull/11) removes the
+  optional `$padding` argument from each of `Zend\Crypt\PublicKey\Rsa`'s
+  `encrypt()` and `decrypt()` methods; you can now specify the value via the
+  `RsaOptions`.
+- [#25](https://github.com/zendframework/zend-crypt/pull/25) removes support for
+  zend-math v2 versions.
+- [#29](https://github.com/zendframework/zend-crypt/pull/29) removes support for
+  PHP 5.5.
+
+### Fixed
+
+- [#22](https://github.com/zendframework/zend-crypt/pull/22) updates all
+  occurrences of `substr()` and `strlen()` to use `mb_substr()` and
+  `mb_strlen()`, respectively. This provides better security with binary values.
+- [#25](https://github.com/zendframework/zend-crypt/pull/25) updates the
+  `Zend\Crypt\Password\Bcrypt` implementation to use `password_hash()` and
+  `password_verify()` internally, as they are supported in all PHP versions we
+  support.
+- [#19](https://github.com/zendframework/zend-crypt/pull/19) fixes the
+  `DiffieHellman` publickey implementation to initialize the `BigInteger`
+  adapter from zend-math as the first operation of its constructor, fixing a
+  fatal error that occurs when binary data is provided.
+
 ## 2.6.1 - TBD
 
 ### Added